# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sandboxd /v1 API
  version: "0.3.0"
  description: >
    The public, tenant-scoped API for sandboxd. This is the only surface a
    console or external integration should use — the internal /sandbox and
    /wake endpoints are not part of this contract. In the local Docker profile,
    requests may use a server session cookie, a bearer API token, or (when
    explicitly disabled) no credential. In the production AKS profile, requests
    use the server session created by OIDC authorization-code PKCE in the
    configured Entra tenant; local passwords and API keys are not production
    authentication mechanisms. Production ownership is keyed to the immutable
    Entra OID and roles are sandboxd.user or sandboxd.admin.
servers:
  - url: http://127.0.0.1:9090
    description: Default loopback bind (docker-compose maps SANDBOXD_API_BIND here)
security:
  - sessionCookieAuth: []
  - bearerAuth: []
  - {} # local auth-disabled mode only
...
paths:
...
  /v1/auth/accounts:
    post:
      operationId: createLocalAccount
      tags: [apps]
      summary: Create a local user account
      description: >
        Available only in the local account profile and requires a sandboxd.admin
        session. New accounts receive the sandboxd.user role only. Passwords are
        write-only and never returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, writeOnly: true }
      responses:
        "201":
          description: Local user account created
          content:
            application/json:
              schema:
                type: object
                required: [id, email, roles]
                properties:
                  id: { type: string }
                  email: { type: string, format: email }
                  roles:
                    type: array
                    items: { type: string }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: sbx_session
      description: >
        HttpOnly, host-only server session cookie. In production it is issued
        after the OIDC authorization-code PKCE flow; the cookie is not a bearer
        credential for preview hosts.
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
...